V0.21.0
major update. (1) remove redundant columns in formule.parse(); cannot remember the reason for this modification, will add comments later once I figure this out; (2) implement 2nd derivative for linear model, a lot of update here;

This version is not stable and is NOT usable. I still update it to git for achieving purpose as too many important and well tested functions are added. In this version, variance parameter tau in reduced model are estimated. This is unnecessary and I will remove those variance parameters from gim model in near future.

There are bugs in this version. Wrong SE appears when reduced model uses variables that are not used in the true model. Those variables could even be transformed from variables that in the true model: e.g. x1 is used in true mode as y ~ x1, while log(x1) is used in reduced model as y ~ log(x1). I will try to figure out the reason and fix it in next version. I am not sure if this could happen for family = 'case-control' or 'binomial'. I am sure that this could happen when family = 'gaussian'.