v0.22.0
major update. This version does not estimate tau, variance parameter in working model. Bugs in v0.21.0 are not fixed in this version. Seems like V.bet, the covariance matrix of external estimate are not correct if family == 'gaussian'. However, using exactly the same true model and working model, but generating data for case-control studies, everything looks right if family == 'case-control'. I update this version to github even if it has potential bugs because I need to test a more efficient version for family=='case-control'.