v0.23.0
major update. (1) Fix the bug mentioned in v0.21.0 and v0.22.0 that for family == 'gaussian', the SE could be quite different from SD. This is because V.bet is not correctly estimated for the external information. Sigma0.lm() is modified. (2) Slightly improve hess.cc(). (3) Update example and built-in dataset for 'gaussian' family. Examples for other families will be added later.