v0.27.0
major update. (1) fix a bug in cov.lm when calculating the SE; (2) change methods for calculating V for lm and lo. The functions for cc is fine; (3) use ... argument to replace niter.

This version is used for simulation in revision. It does not work if formula is specified as y~.