v0.32.0
major update. This version is not fully tested but seems be fine. (1)
allow for reference specified for external data under case-control
studies. All the derivation are similar to the one without reference;
(2) much better support for formula input, for example, ‘y~.’. This is
done by store fitted model with internal and external formula rather
than passing those formulas to init functions. The former solution is
not that good because data (and ref) may have been argued during data
preprocess and formula parsing, in that case formula like ‘y~.’ behaves
very differently with data of different variables in it. (3) fix a
minor bug when calculate covariance for case-control data (vcov.cc)