v0.32.2
Unstable version. (1) fix a bug when alp is empty, i.e. even intercept
is provided. Many functions are modified. Need comprehensive simulation
to check no new bug is added by this modification. (2) fix a “bug” (not
sure) when bet fitted in a single model are given as two elements in
model. The result is supposed to be the same as if bet are given in a
single element in model. However, this is not the case in previous
version. Might be complicated to figure out the reason. Used a simple
way to fix this, i.e. collapse models if multiple models could be
merged into a single element in “model”. (2) implement three functions
compute.*** to make NR more readable. (3) some other functions,
including check.***.R, convex**.R, **loop**.R, wolfe**.R are used to
implement two-loop algorithm that fails. Very likely to be removed in
future stable version.