v0.34.1
Kai found that the covariance matrix, V, of external summary data is
not updated during the while loop in the main function. This is not a
bug because V is still a consistent estimate. This update is temporary
as it breaks the nice structure of the function gim.default. See the
newly added lines 60-61: it is hard to understand why setting pr0 and
Delta to be NULL. This is also the reason why I chose to not update pr0
and Delta in former versions. With very limited testing on this updated
version, I did not see unusual result, but more comprehensive testing
is needed before releasing to CRAN. This modification only apply for
the case where ref == NULL, i.e external and internal data are from the
same population.